Remove unused libraries
