--- detekt-rules/src/main/resources/META-INF/services/io.gitlab.arturbosch.detekt.api.RuleSetProvider
@@ -1,1 +1,0 @@
-br.com.wesjon.linters.detekt.CustomRuleProvider
