+++ sample/src/main/java/br/com/wesjon/MainActivity.kt
@@ -1,18 +1,5 @@
 package br.com.wesjon
 
-import android.os.Bundle
-import com.google.android.material.bottomnavigation.BottomNavigationView
 import androidx.appcompat.app.AppCompatActivity
-import androidx.navigation.findNavController
-import androidx.navigation.ui.setupWithNavController
 
-class MainActivity : AppCompatActivity(R.layout.activity_main) {
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        val navView: BottomNavigationView = findViewById(R.id.nav_view)
-
-        val navController = findNavController(R.id.nav_host_fragment)
-        navView.setupWithNavController(navController)
-    }
-}
+class MainActivity : AppCompatActivity(R.layout.activity_main)
