--- sample/src/main/java/br/com/wesjon/ui/dashboard/DashboardFragment.kt
@@ -1,22 +1,0 @@
-package br.com.wesjon.ui.dashboard
-
-import android.os.Bundle
-import android.view.View
-import android.widget.TextView
-import androidx.fragment.app.Fragment
-import androidx.fragment.app.viewModels
-import androidx.lifecycle.Observer
-import br.com.wesjon.R
-
-class DashboardFragment : Fragment(R.layout.fragment_dashboard) {
-
-    private val dashboardViewModel: DashboardViewModel by viewModels()
-
-    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
-        super.onViewCreated(view, savedInstanceState)
-        val textView: TextView = view.findViewById(R.id.text_dashboard)
-        dashboardViewModel.text.observe(viewLifecycleOwner, Observer {
-            textView.text = it
-        })
-    }
-}
