--- sample/src/main/java/br/com/wesjon/ui/home/HomeFragment.kt
@@ -1,23 +1,0 @@
-package br.com.wesjon.ui.home
-
-import android.os.Bundle
-import android.view.View
-import android.widget.TextView
-import androidx.fragment.app.Fragment
-import androidx.fragment.app.viewModels
-import androidx.lifecycle.Observer
-import br.com.wesjon.R
-
-class HomeFragment : Fragment(R.layout.fragment_home) {
-
-    private val homeViewModel: HomeViewModel by viewModels()
-
-    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
-        super.onViewCreated(view, savedInstanceState)
-
-        val textView: TextView = view.findViewById(R.id.text_home)
-        homeViewModel.text.observe(viewLifecycleOwner, Observer {
-            textView.text = it
-        })
-    }
-}
