--- sample/src/main/java/br/com/wesjon/ui/notifications/NotificationsFragment.kt
@@ -1,23 +1,0 @@
-package br.com.wesjon.ui.notifications
-
-import android.os.Bundle
-import android.view.View
-import android.widget.TextView
-import androidx.fragment.app.Fragment
-import androidx.fragment.app.viewModels
-import androidx.lifecycle.Observer
-import br.com.wesjon.R
-
-class NotificationsFragment : Fragment(R.layout.fragment_notifications) {
-
-    private val notificationsViewModel: NotificationsViewModel by viewModels()
-
-    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
-        super.onViewCreated(view, savedInstanceState)
-
-        val textView: TextView = view.findViewById(R.id.text_notifications)
-        notificationsViewModel.text.observe(viewLifecycleOwner, Observer {
-            textView.text = it
-        })
-    }
-}
