--- sample/src/test/java/br/com/wesjon/ExampleUnitTest.kt
@@ -1,16 +1,0 @@
-package br.com.wesjon
-
-import org.junit.Assert.assertEquals
-import org.junit.Test
-
-/**
- * Example local unit test, which will execute on the development machine (host).
- *
- * See [testing documentation](http://d.android.com/tools/testing).
- */
-class ExampleUnitTest {
-    @Test
-    fun `addition is correct`() {
-        assertEquals(4, 2 + 2)
-    }
-}
