+++ settings.gradle
@@ -1,3 +1,3 @@
 include ':sample'
-include ':detekt-rules'
-rootProject.name = "Android Linters"
+include ':android-rules'
+rootProject.name = "detekt-rules-android"
